REFACTOR: 유저 정보 전솔 api 분리
