+++ src/screens/account/Getprofile.js
@@ -2,50 +2,17 @@ import { useNavigation } from '@react-navigation/native';
 import { useState } from 'react';
 import { View, Text, TouchableOpacity, StyleSheet, TextInput } from 'react-native';
 import ProfileImage from '../../components/getProfile/profileImage';
-import getToken from '../../server/getToken';
-import { SERVER_HOST } from "@env";
-import axios from 'axios';
-import { Platform } from 'react-native';
-
+import postUserInfo from '../../server/getInfo/post-profile';
 
 // test용 스크린
 export default function GetProfile() {
-    const navigation = useNavigation();
-    const token = getToken();
-
+  const navigation = useNavigation();
     const [nickname, setNickname] = useState("");
     const [image, setImage] = useState([]);
 
-	const postUserInfo = async () => {
-    const formData = new FormData();
-    // 이미지 용량이 너무 커서 압축이 필요하다.
-    // 안드는 jpeg, ios는 jpg로 고정
-    const file = {
-      name: image[0],
-      type: Platform.OS === 'android' ? 'image/jpeg' : 'image/jpg',
-      uri: image[2],
-    }
-
-    formData.append('profileImage', file);
-    formData.append('nickname', nickname);
-
-		const config = {
-			headers: {
-        'Content-Type': 'multipart/form-data',
-        'Authorization': `Bearer ${token}`,
-			}
-		};
-		
-		try {
-		  const response = await axios.post(`${SERVER_HOST}/api/v1/auth/user-info`, formData, config);
-      if(response.data.success == true) {
-        navigation.navigate('Daily');  // 우선 daliy로 이동
-      } else {
-        alert('에러가 발생했습니다. 이미지를 다시 보내주세요.');
-      }
-		} catch (error) {
-		  console.log('에러가 있습니다. ', error);
-		}
+    // 유저 정보 전송
+	const clickPost = async () => {
+    postUserInfo(image, nickname, navigation);
 	  };
 
     return (
@@ -59,7 +26,7 @@ export default function GetProfile() {
         placeholder='닉네임을 입력해주세요'
         placeholderTextColor='#ffffff' />
     </View>
-    <TouchableOpacity style={styles.button} onPress={postUserInfo}>
+    <TouchableOpacity style={styles.button} onPress={clickPost}>
         <Text style={styles.buttontext}>확인</Text>
     </TouchableOpacity>
       </View>
